requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
packaging==21.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.0
